requirements.txt
accelerate==0.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
av==13.1.0
beautifulsoup4==4.14.2
bitsandbytes==0.48.1
blinker==1.9.0
blis==0.7.11
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloup==3.0.8
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.11
datasets==4.2.0
decorator==5.2.1
deprecation==2.1.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
fastapi==0.120.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gguf==0.17.1
glcontext==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gpt4all==2.8.2
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
isosurfaces==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
llama_cpp_python==0.3.16
manim==0.19.0
ManimPango==0.6.1
mapbox_earcut==1.0.3
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.13
networkx==3.5
numpy==2.3.4
openai==1.55.3
packaging==25.0
pandas==2.2.3
pathlib_abc==0.1.1
pathy==0.11.0
peft==0.7.1
pillow==12.0.0
pluggy==1.6.0
postgrest==2.24.0
preshed==3.0.10
prometheus_client==0.20.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
pyglet==2.1.10
pyglm==2.8.2
Pygments==2.19.2
PyJWT==2.10.1
pyllamacpp==2.4.3
PyMuPDF==1.26.6
pyparsing==3.2.5
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
realtime==2.24.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.3
screeninfo==0.8.1
seaborn==0.13.2
sentencepiece==0.2.1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
skia-pathops==0.8.0.post2
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.8
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
srt==3.5.3
starlette==0.49.1
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
svgelements==1.9.6
sympy==1.13.1
thinc==8.1.12
tokenizers==0.22.1
torch==2.6.0+cpu
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@60b68e304cf4b6569b0660a13b558b929d4b0e77
trl==0.7.4
typeguard==4.4.4
typer==0.15.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.9.35
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
xxhash==3.6.0
yarl==1.22.0
